Alterado estrutura de como funciona o STACK máximo de itens no storage.
Configuração 'storage_auto_new_entry' adicionada.

Quando 'storage_auto_new_entry' for ligada (true/1), novas entradas (de itens que podem ser acumulados, ex: Poções) terão seu modo de entrada um pouco alterados.

A Forma anterior era, se você quiser inserir uma quantidade de itens que supere a configuração de 'MAX_AMOUNT' (common/mmo.h) ele não permitiria.
Este modo deve ser mantido caso se a configuração não estiver ativada.

No novo modo, quando o limite máximo for atingido, uma nova entrada no storage será criada. (Claro, respeitando 'MAX_STORAGE')

+++ src/common/mmo.h
@@ -256,7 +256,7 @@ struct quest {
 struct item {
 	int id;
 	short nameid;
-	short amount;
+	int amount;
 	unsigned int equip; // Location(s) where item is equipped (using enum equip_pos for bitmasking).
 	char identify;
 	char refine;
